+++ RELEASE_NOTES.md
@@ -16,6 +16,8 @@ More details at <a href="https://nighttime-imaging.eu/donate/" target="_blank">n
 ## Bugfixes
 - The profile chooser on startup did not remember the on/off selection to save the selected profile
 - Fixed SVBony Native driver, that was sometimes showing the exposure before the latest one after a cancelled exposure
+- Fixed PlayerOne resolution not fully resetting to complete size after subframe or binning
+- Added automatic retry of exposure start for PlayerOne driver when POA_ERROR_EXPOSING error happens
 
 # Version 2.2